// src/types/GamificationData.ts
export interface GamificationData {
  id?: string;
  userId: string;
  xp: number;
  coins: number;
  xp30d: number;
  vitality: number;
  mood: string;
  xpMultiplier: number;
  xpMultiplierExpiry: number;
  str: number;
  int: number;
  cre: number;
  soc: number;
  aspect: string;
  rankIdx: number;
  rankTier: string;
  rankDiv: number;
}